+++ src/fixture/clubfixture.py
@@ -34,7 +34,7 @@ class ClubFactory(factory.django.DjangoModelFactory):
 
     @factory.post_generation
     def users(self, create, extracted, **kwargs):
-        if not create:
+        if not create:  # pragma: no use
             return
 
         if extracted:
