+++ src/fixture/topicfixture.py
@@ -20,7 +20,7 @@ class TopicFactory(factory.django.DjangoModelFactory):
 
     @factory.post_generation
     def users(self, create, extracted, **kwargs):
-        if not create:
+        if not create:  # pragma: no use
             return
 
         if extracted:
