+++ src/gallery/tests.py
@@ -1,0 +1,1 @@
+"""No test"""
