+++ src/test/__init__.py
@@ -1,0 +1,2 @@
+# TODO: make sure travis doesn't fail
+# TODO: make installed app differ for fixture app as it was previous
