--- src/test/test_gallery.py
@@ -1,1 +1,0 @@
-"""No test"""
